Create README.md
This project is aimed to developing an online park ticket booking system. The entire project has been developed keeping in view of the distributed client server computing technology, in mind. The admin is to create an e-Information about the park booking and admin is to  schedule the park booking that are helpful to peoples for their convinience. Through this application any peoples who is intrest to spent time in a park  can register himself in the same way.,and also able to direct chat with admin through this application .Admin is the main authority who can do addition, deletion, and modification if required.